Extract ITIR visitor to separate class
