+++ src/gt4py/next/program_processors/runners/dace_fieldview/gtir_builtins/gtir_builtin_field_operator.py
@@ -13,50 +13,25 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 
-from dataclasses import dataclass
 from typing import Callable, TypeAlias
 
 import dace
 
-from gt4py import eve
 from gt4py.next.common import Connectivity, Dimension
 from gt4py.next.iterator import ir as itir
 from gt4py.next.iterator.ir_utils import common_pattern_matcher as cpm
 from gt4py.next.program_processors.runners.dace_fieldview.gtir_builtins.gtir_builtin_translator import (
     GTIRBuiltinTranslator,
 )
-from gt4py.next.program_processors.runners.dace_fieldview.gtir_types import MATH_BUILTINS_MAPPING
-from gt4py.next.program_processors.runners.dace_fieldview.utility import (
-    as_dace_type,
-    get_domain,
-    unique_name,
+from gt4py.next.program_processors.runners.dace_fieldview.gtir_to_tasklet import (
+    GTIRToTasklet,
+    ValueExpr,
 )
+from gt4py.next.program_processors.runners.dace_fieldview.utility import get_domain, unique_name
 from gt4py.next.type_system import type_specifications as ts
 
 
-@dataclass(frozen=True)
-class LiteralExpr:
-    """Any symbolic expression that can be evaluated at compile time."""
-
-    value: dace.symbolic.SymbolicType
-
-
-@dataclass(frozen=True)
-class SymbolExpr:
-    """The data access to a scalar or field through a symbolic reference."""
-
-    data: str
-
-
-@dataclass(frozen=True)
-class ValueExpr:
-    """The result of a computation provided by a tasklet node."""
-
-    node: dace.nodes.Tasklet
-    connector: str
-
-
-class GTIRBuiltinAsFieldOp(GTIRBuiltinTranslator, eve.NodeVisitor):
+class GTIRBuiltinAsFieldOp(GTIRBuiltinTranslator):
     """Generates the dataflow subgraph for the `as_field_op` builtin function."""
 
     TaskletConnector: TypeAlias = tuple[dace.nodes.Tasklet, str]
@@ -65,7 +40,6 @@ class GTIRBuiltinAsFieldOp(GTIRBuiltinTranslator, eve.NodeVisitor):
     stencil_args: list[Callable]
     field_domain: dict[Dimension, tuple[dace.symbolic.SymbolicType, dace.symbolic.SymbolicType]]
     field_type: ts.FieldType
-    input_connections: list[TaskletConnector]
     offset_provider: dict[str, Connectivity | Dimension]
 
     def __init__(
@@ -77,7 +51,6 @@ class GTIRBuiltinAsFieldOp(GTIRBuiltinTranslator, eve.NodeVisitor):
         offset_provider: dict[str, Connectivity | Dimension],
     ):
         super().__init__(sdfg, state)
-        self.input_connections = []
         self.offset_provider = offset_provider
 
         assert cpm.is_call_to(node.fun, "as_fieldop")
@@ -102,11 +75,10 @@ class GTIRBuiltinAsFieldOp(GTIRBuiltinTranslator, eve.NodeVisitor):
         self.stencil_args = stencil_args
 
     def build(self) -> list[tuple[dace.nodes.Node, ts.FieldType | ts.ScalarType]]:
-        assert len(self.input_connections) == 0
-
         # generate a tasklet node implementing the stencil function and represent
         # the field operator as a mapped tasklet, which will range over the field domain
-        output_expr = self.visit(self.stencil_expr.expr)
+        taskgen = GTIRToTasklet(self.sdfg, self.head_state, self.offset_provider)
+        input_connections, output_expr = taskgen.visit(self.stencil_expr)
         assert isinstance(output_expr, ValueExpr)
 
         # allocate local temporary storage for the result field
@@ -150,13 +122,13 @@ class GTIRBuiltinAsFieldOp(GTIRBuiltinTranslator, eve.NodeVisitor):
         map_ranges = {f"i_{dim.value}": f"{lb}:{ub}" for dim, (lb, ub) in self.field_domain.items()}
         me, mx = self.head_state.add_map(unique_name("map"), map_ranges)
 
-        for (input_node, input_connector), input_param in self.input_connections:
+        for input_expr, input_param in input_connections:
             assert input_param in data_nodes
             self.head_state.add_memlet_path(
                 data_nodes[input_param],
                 me,
-                input_node,
-                dst_conn=input_connector,
+                input_expr.node,
+                dst_conn=input_expr.connector,
                 memlet=input_memlets[input_param],
             )
         self.head_state.add_memlet_path(
@@ -164,80 +136,3 @@ class GTIRBuiltinAsFieldOp(GTIRBuiltinTranslator, eve.NodeVisitor):
         )
 
         return [(field_node, self.field_type)]
-
-    def _visit_shift(self, node: itir.FunCall) -> str:
-        raise NotImplementedError
-
-    def visit_FunCall(self, node: itir.FunCall) -> ValueExpr | SymbolExpr:
-        inp_tasklets = {}
-        inp_symbols = set()
-        inp_connectors = set()
-        node_internals = []
-        for i, arg in enumerate(node.args):
-            arg_expr = self.visit(arg)
-            if isinstance(arg_expr, LiteralExpr):
-                # use the value without adding any connector
-                node_internals.append(arg_expr.value)
-            else:
-                if isinstance(arg_expr, ValueExpr):
-                    # the value is the result of a tasklet node
-                    connector = f"__inp_{i}"
-                    inp_tasklets[connector] = arg_expr
-                else:
-                    # the value is the result of a tasklet node
-                    assert isinstance(arg_expr, SymbolExpr)
-                    connector = f"__inp_{arg_expr.data}"
-                    inp_symbols.add((connector, arg_expr.data))
-                inp_connectors.add(connector)
-                node_internals.append(connector)
-
-        if cpm.is_call_to(node, "deref"):
-            assert len(inp_tasklets) == 0
-            assert len(inp_symbols) == 1
-            _, data = inp_symbols.pop()
-            return SymbolExpr(data)
-
-        elif cpm.is_call_to(node.fun, "shift"):
-            code = self._visit_shift(node.fun)
-
-        elif isinstance(node.fun, itir.SymRef):
-            # create a tasklet node implementing the builtin function
-            builtin_name = str(node.fun.id)
-            if builtin_name in MATH_BUILTINS_MAPPING:
-                fmt = MATH_BUILTINS_MAPPING[builtin_name]
-                code = fmt.format(*node_internals)
-            else:
-                raise NotImplementedError(f"'{builtin_name}' not implemented.")
-
-            out_connector = "__out"
-            tasklet_node = self.head_state.add_tasklet(
-                unique_name("tasklet"),
-                inp_connectors,
-                {out_connector},
-                "{} = {}".format(out_connector, code),
-            )
-
-        else:
-            raise NotImplementedError(f"Unexpected 'FunCall' node ({node}).")
-
-        for input_conn, inp_expr in inp_tasklets.items():
-            self.head_state.add_edge(
-                inp_expr.node, inp_expr.connector, tasklet_node, input_conn, dace.Memlet()
-            )
-        self.input_connections.extend(
-            ((tasklet_node, connector), data) for connector, data in inp_symbols
-        )
-        return ValueExpr(tasklet_node, out_connector)
-
-    def visit_Literal(self, node: itir.Literal) -> LiteralExpr:
-        cast_sym = str(as_dace_type(node.type))
-        cast_fmt = MATH_BUILTINS_MAPPING[cast_sym]
-        typed_value = cast_fmt.format(node.value)
-        return LiteralExpr(typed_value)
-
-    def visit_SymRef(self, node: itir.SymRef) -> SymbolExpr:
-        """
-        Symbol references are mapped to tasklet connectors that access some kind of data.
-        """
-        sym_name = str(node.id)
-        return SymbolExpr(sym_name)
